pkg/util: move trace.go to k8s.io/apiserver/pkg/util/trace
